Feature bet records

See merge request Android/sport-android!237

+++ app/src/main/java/org/cxct/sportlottery/network/bet/list/Other.kt
@@ -8,7 +8,7 @@ import org.cxct.sportlottery.network.bet.MatchOdd
 @JsonClass(generateAdapter = true)
 data class Other (
     @Json(name = "totalAmount")
-    val totalAmount: Int?,
+    val totalAmount: Long?,
     @Json(name = "win")
     val win: Double?,
     @Json(name = "valueBetAmount")
